fix: Attachment from slack to chatwoot
